refactoring : DTO 및 엔티티 분류
