+++ src/main/java/com/example/bcsd/Controller/ArticleController.java
@@ -1,4 +1,4 @@
-package com.example.bcsd;
+package com.example.bcsd.Controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -8,15 +8,18 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
 
+import com.example.bcsd.DTO.ArticleRequestDTO;
+import com.example.bcsd.DTO.ArticleEntity;
+
 @RestController
 public class ArticleController {
 
-    private final Map<Long, Article> articles = new HashMap<>();
+    private final Map<Long, ArticleEntity> articles = new HashMap<>();
     private final AtomicLong idCount = new AtomicLong();
 
     @GetMapping("/article/{id}")
-    public ResponseEntity<Article> getArticle(@PathVariable Long id) {
-        Article article = articles.get(id);
+    public ResponseEntity<ArticleEntity> getArticle(@PathVariable Long id) {
+        ArticleEntity article = articles.get(id);
 
         if (article == null) {
             return ResponseEntity.notFound().build();
@@ -26,21 +29,21 @@ public class ArticleController {
     }
 
     @PostMapping("/article")
-    public ResponseEntity<Article> postArticle(@RequestBody ArticleRequestDTO request) {
+    public ResponseEntity<ArticleEntity> postArticle(@RequestBody ArticleRequestDTO request) {
         long newId = idCount.incrementAndGet();
-        Article newArticle = new Article(newId, request.description());
+        ArticleEntity newArticle = new ArticleEntity(newId, request.description());
         articles.put(newId, newArticle);
 
         return ResponseEntity.status(HttpStatus.CREATED).body(newArticle);
     }
 
     @PutMapping("/article/{id}")
-    public ResponseEntity<Article> putArticle(@PathVariable Long id, @RequestBody ArticleRequestDTO request) {
+    public ResponseEntity<ArticleEntity> putArticle(@PathVariable Long id, @RequestBody ArticleRequestDTO request) {
         if (!articles.containsKey(id)) {
             return ResponseEntity.notFound().build();
         }
 
-        Article updateArticle = new Article(id, request.description());
+        ArticleEntity updateArticle = new ArticleEntity(id, request.description());
         articles.put(id, updateArticle);
 
         return ResponseEntity.ok(updateArticle);
@@ -48,18 +51,12 @@ public class ArticleController {
 
     @DeleteMapping("/article/{id}")
     public ResponseEntity<Void> deleteArticle(@PathVariable Long id) {
-        Article removedArticle = articles.remove(id);
+        ArticleEntity removedArticle = articles.remove(id);
 
         if (removedArticle == null) {
             return ResponseEntity.notFound().build();
         }
 
         return ResponseEntity.noContent().build();
-    }
-
-    public record Article(Long id, String description) {
-    }
-
-    public record ArticleRequestDTO(String description) {
     }
 }
