+++ src/main/java/com/example/bcsd/Controller/IntroduceController.java
@@ -1,4 +1,4 @@
-package com.example.bcsd;
+package com.example.bcsd.Controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
