+++ src/main/java/com/example/bcsd/DTO/ArticleEntity.java
@@ -1,0 +1,6 @@
+package com.example.bcsd.DTO;
+
+public record ArticleEntity(
+        Long id,
+        String description) {
+}
