+++ src/main/java/com/example/bcsd/DTO/ArticleRequestDTO.java
@@ -1,0 +1,4 @@
+package com.example.bcsd.DTO;
+
+public record ArticleRequestDTO(String description) {
+}
